pyproject.toml
[project]
name = "urpe-agent-mvp"
version = "0.1.0"
description = "CLI AI agent with Typer + Rich, LiteLLM, and SQLite memory"
authors = [
    { name = "Bruno", email = "bruno@example.com" } # Placeholder email
]
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "litellm>=1.0.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
urpe = "urpe.cli:app"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/urpe"]
